third_party/blink/web_tests/fast/forms/pseudo-open.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/whatwg/html/pull/10126">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="./resources/picker-common.js"></script>

<div class=supported>
  <input type=date>
  <input type=datetime-local>
  <input type=week>
  <input type=month>
  <input type=time>
  <input type=color>
</div>

<input type=text list=datalist>
<datalist id=datalist>
  <option>one</option>
  <option>two</option>
</datalist>

<script>
document.querySelectorAll('.supported > input').forEach(input => {
  const inputType = input.getAttribute('type');
  promise_test(async () => {
    document.body.appendChild(input);
    assert_false(input.matches(':open'),
      'Should not match :open at the start of the test.');

    await openPicker(input);
    assert_true(input.matches(':open'),
      'Should match :open after opening the picker.');

    await attemptToClosePicker(input);
    assert_false(input.matches(':open'),
      'Should not match :open after closing the picker.');
  }, `CSS :open for <input type=${inputType}>`);
});

// TODO(crbug.com/383306186): Support :open for <input type=text list=datalist>
promise_test(async () => {
  const input = document.querySelector('input[list]');
  await test_driver.click(input);
  await new Promise(requestAnimationFrame);
  await new Promise(requestAnimationFrame);
  assert_false(input.matches(':open'),
    ':open is not supported yet.');
}, 'CSS :open for <input type=text list=datalist>');
</script>
